Restructure robust_problem.py by moving SolverStats and torch expression generation to separate files
